# frontend/moderator-dashboard-ui/docs/verification/backend/openapi_verified.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "B2B Platform API",
    "description": "API for B2B Platform - supplier moderation and parsing system",
    "version": "1.0.0"
  },
  "paths": {
...
    "/parsing/start": {
      "post": {
        "tags": [
          "Parsing"
        ],
        "summary": "Start Parsing Endpoint",
        "description": "Start parsing for a keyword.",
        "operationId": "start_parsing_endpoint_parsing_start_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StartParsingRequestDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/parsing/status/{run_id}": {
      "put": {
        "tags": [
          "Parsing"
        ],
        "summary": "Update Parsing Status Endpoint",
        "description": "Update parsing run status (for CAPTCHA and progress updates).",
        "operationId": "update_parsing_status_endpoint_parsing_status__run_id__put",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": true,
                "default": {},
                "title": "Request Data"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Parsing"
        ],
        "summary": "Get Parsing Status Endpoint",
        "description": "Get parsing status by run ID.",
        "operationId": "get_parsing_status_endpoint_parsing_status__run_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ParsingStatusResponseDTO"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ParsingStatusResponseDTO": {
        "properties": {
          "run_id": {
            "type": "string",
            "title": "Run Id"
          },
          "keyword": {
            "type": "string",
            "title": "Keyword"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "started_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "finished_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          },
          "error_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Message"
          },
          "resultsCount": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resultscount"
          }
        },
        "type": "object",
        "required": [
          "run_id",
          "keyword",
          "status"
        ],
        "title": "ParsingStatusResponseDTO",
        "description": "Response DTO for parsing status."
      },
...
      "StartParsingRequestDTO": {
        "properties": {
          "keyword": {
            "type": "string",
            "title": "Keyword"
          },
          "depth": {
            "type": "integer",
            "title": "Depth",
            "description": "Number of search result pages to parse (depth)",
            "default": 50
          },
          "source": {
            "type": "string",
            "title": "Source",
            "description": "Source for parsing: 'google', 'yandex', or 'both'",
            "default": "google"
          }
        },
        "type": "object",
        "required": [
          "keyword"
        ],
        "title": "StartParsingRequestDTO",
        "description": "Request DTO for starting parsing."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}